Use a copy of global Schema fields for serialization.
The problem is with Schema's setup, which is done on the class level,
so Schema and it's fields are global objects, initialized during the
application startup.

The global field objects are available as SchemaClass._declared_fields
or as schema_object._declared_fields when we are working with the Schema
instance.

Schema itself does this:

```
    self.declared_fields = copy.deepcopy(self._declared_fields)`
```

before doing anything with the fields.

So `self._declared_fields` are global, and `self.declared_fields` are
local the field object.

This issue was tracked down with `Nested` field, which has `__schema`
property.
This property is not set on object creation and modified later, during
the serialization (see https://github.com/marshmallow-code/marshmallow/blob/4457426e7447195e5b390605ec90349b81ded507/marshmallow/fields.py#L398).

So the following assertion was always true:

```
    assert GalleryLocationSchema._declared_fields['city']._Nested__schema is None
```

But it failed after exporting APISpec to json:

```
apispec.init_app(current_app)
data = apispec.to_dict()
json.dumps(data)
assert GalleryLocationSchema._declared_fields['city']._Nested__schema is None  # Fails!
```

The assert started failing after json.dupms, which in turn triggers
serialization of the data in `LazyDict`s inside the
apispec.ext.marshmallow.swagger which triggered the serialization
process on global `Schema._declared_fields`.

The fix is to do the same thing `Schema.__init__` does and copy the
fields before doing anything to them.